修复了数据量过大parseInt解析失败的问题

+++ src/main/java/com/zcq/pojo/SwitchInfo.java
@@ -28,9 +28,9 @@ public class SwitchInfo {
     /* Redis解析填充 */
     private int averagepkts;
     private int maxpkts;
-    private int averagebytes;
-    private int maxbytes;
-    private int totalbytes;
+    private long averagebytes;
+    private long maxbytes;
+    private long totalbytes;
     private int totalflownums;
     private int totalflowrate;
 }
